pyproject.toml
[project]
name = "gt-address-parser"
version = "0.1.0"
description = "Sistema de parsing e inteligencia de direcciones guatemaltecas (NER con Bi-LSTM)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "openai>=2.24.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "pytorch-crf>=0.7.2",
    "rich>=14.3.3",
    "seqeval>=1.2.2",
    "torch>=2.10.0",
    "typer>=0.24.1",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.4",
]

[project.scripts]
generate-data = "src.data_gen.generator:app"
preprocess    = "src.data_gen.preprocess:app"
train-ner     = "src.model.train:app"
eval-report   = "src.model.eval_report:app"
api-server    = "src.api.app:serve"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
